+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 ------------------
 
 - `LayerSegmentsImageSegmentationReader` now suggest using `--lenient` flag in exception in case image not binary
+- added the `discard-by-name` filter that allows user to discard images based on name, either exact match of regexp
+  (matching sense can be inverted)
 
 
 0.0.5 (2025-01-13)