+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - `LayerSegmentsImageSegmentationReader` now suggest using `--lenient` flag in exception in case image not binary
 - added the `discard-by-name` filter that allows user to discard images based on name, either exact match of regexp
   (matching sense can be inverted)
+- requiring seppl>=0.2.9 now
 
 
 0.0.5 (2025-01-13)