+++ CHANGES.rst
@@ -13,6 +13,7 @@ Changelog
 - added the `generate_palette_list` method to `idc.api` which turns a predefined palette name or comma-separated
   list of RGB values into a flat list of int values, e.g., used for indexed PNG files
 - exposed method `save_image` through `idc.api`
+- `filter-labels` now handles not specifying any labels and only regexp
 
 
 0.0.5 (2025-01-13)